Fix for upload, added proper permissions for credits and tokens, normalised view sections js and table margins in user.php, a little bit more on bonus shop section

+++ sections/user/takemoderate.php
@@ -71,6 +71,10 @@ $FLTokens = isset($_POST['FLTokens'])?$_POST['FLTokens'] : 0;
 if(!is_number($FLTokens)) {
 	error(0);
 }
+$BonusCredits = isset($_POST['BonusCredits'])?$_POST['BonusCredits'] : 0;
+if(!is_number($BonusCredits)) {
+	error(0);
+}
 
 $WarnLength = (int)$_POST['WarnLength'];
 $ExtendWarning = (int)$_POST['ExtendWarning'];
@@ -141,7 +145,8 @@ $DB->query("SELECT
 	m.RequiredRatio,
 	m.FLTokens,
 	i.RatioWatchEnds,
-	SHA1(i.AdminComment) AS CommentHash
+	SHA1(i.AdminComment) AS CommentHash,
+	m.Credits
 	FROM users_main AS m
 	JOIN users_info AS i ON i.UserID = m.ID
 	LEFT JOIN permissions AS p ON p.ID=m.PermissionID
@@ -312,14 +317,14 @@ if ($Downloaded!=$Cur['Downloaded'] && $Downloaded!=$_POST['OldDownloaded'] && (
 	$Cache->delete_value('users_stats_'.$UserID);
 }
 */
-if ($AdjustUpValue != 0 && (check_perms('users_edit_ratio') 
-                        || (check_perms('users_edit_own_ratio') && $UserID == $LoggedUser['ID']))){
+if ($AdjustUpValue != 0 && ((check_perms('users_edit_ratio') && $UserID != $LoggedUser['ID'])
+                        || (check_perms('users_edit_own_ratio') && $UserID == $LoggedUser['ID'])) ){
       $Uploaded = $Cur['Uploaded'] + $AdjustUpValue;
 	$UpdateSet[]="Uploaded='".$Uploaded."'";
 	$EditSummary[]="uploaded changed from ".get_size($Cur['Uploaded'])." to ".get_size($Uploaded);
 	$Cache->delete_value('users_stats_'.$UserID);
 }
-if ($AdjustDownValue != 0 && (check_perms('users_edit_ratio') 
+if ($AdjustDownValue != 0 && ((check_perms('users_edit_ratio') && $UserID != $LoggedUser['ID'])
                         || (check_perms('users_edit_own_ratio') && $UserID == $LoggedUser['ID']))){
       $Downloaded = $Cur['Downloaded'] + $AdjustDownValue;
 	$UpdateSet[]="Downloaded='".$Downloaded."'";
@@ -327,12 +332,20 @@ if ($AdjustDownValue != 0 && (check_perms('users_edit_ratio')
 	$Cache->delete_value('users_stats_'.$UserID);
 }
 
-
-if ($FLTokens!=$Cur['FLTokens'] && (check_perms('users_edit_ratio') || (check_perms('users_edit_own_ratio') && $UserID == $LoggedUser['ID']))) {
+if ($FLTokens!=$Cur['FLTokens'] && ((check_perms('users_edit_tokens')  && $UserID != $LoggedUser['ID'])
+                        || (check_perms('users_edit_own_tokens') && $UserID == $LoggedUser['ID']))) {
 	$UpdateSet[]="FLTokens=".$FLTokens;
 	$EditSummary[]="Freeleech Tokens changed from ".$Cur['FLTokens']." to ".$FLTokens;
 	$HeavyUpdates['FLTokens'] = $FLTokens;
 }
+
+if ($BonusCredits!=$Cur['Credits'] && ((check_perms('users_edit_credits') && $UserID != $LoggedUser['ID']) 
+                        || (check_perms('users_edit_own_credits') && $UserID == $LoggedUser['ID']))) {
+	$UpdateSet[]="Credits=".$BonusCredits;
+	$EditSummary[]="Bonus Credits changed from ".$Cur['Credits']." to ".$BonusCredits;
+	$HeavyUpdates['Credits'] = $BonusCredits;
+}
+
 
 if ($Invites!=$Cur['Invites'] && check_perms('users_edit_invites')) {
 	$UpdateSet[]="invites='$Invites'";
